requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.10.0
django-filter==21.1
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
et-xmlfile==1.1.0
Jinja2==3.1.1
Markdown==3.3.4
MarkupSafe==2.1.1
numpy==1.21.2
openpyxl==3.0.9
pandas==1.3.3
psycopg2==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
wfastcgi==3.0.0
XlsxWriter==3.0.2